# app-barbearia-rascunho/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: API MVP Barbearia
  version: 1.0.0
  description: Contrato inicial da API para o MVP do aplicativo de barbearia.
servers:
  - url: https://api.exemplo.com/api/v1
    description: Producao
  - url: http://localhost:3000/api/v1
    description: Desenvolvimento
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Criar conta de cliente
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Conta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/login:
    post:
      tags: [Auth]
      summary: Autenticar usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login realizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Solicitar recuperacao de senha
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Instrucao enviada se o e-mail existir
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      tags: [Auth]
      summary: Retornar usuario autenticado
      responses:
        "200":
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /barbers:
    get:
      tags: [Barbers]
      summary: Listar barbeiros
      responses:
        "200":
          description: Lista de barbeiros
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Barber"
    post:
      tags: [Barbers]
      summary: Criar barbeiro
      description: Apenas admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBarberRequest"
      responses:
        "201":
          description: Barbeiro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberResponse"
  /barbers/{id}:
    get:
      tags: [Barbers]
      summary: Buscar barbeiro
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Barbeiro encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberResponse"
    patch:
      tags: [Barbers]
      summary: Editar barbeiro
      description: Apenas admin.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBarberRequest"
      responses:
        "200":
          description: Barbeiro atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberResponse"
    delete:
      tags: [Barbers]
      summary: Desativar barbeiro
      description: Apenas admin.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Barbeiro desativado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarberResponse"
...
  /bookings/barber/me:
    get:
      tags: [Bookings]
      summary: Listar agenda do barbeiro logado
      description: Apenas barbeiro.
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Agenda do barbeiro
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Booking"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Nao autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflito
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Dados invalidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UserRole:
      type: string
      enum: [client, barber, admin]
    BookingStatus:
      type: string
      enum: [scheduled, confirmed, completed, canceled, no_show]
    RegisterRequest:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string
        message:
          type: string
    UserResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/User"
        message:
          type: string
...
    CreateBarberRequest:
      type: object
      required: [name, email, phone, password, publicName]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
        publicName:
          type: string
        specialty:
          type: string
        isActive:
          type: boolean
    UpdateBarberRequest:
      type: object
      properties:
        publicName:
          type: string
        specialty:
          type: string
        photoUrl:
          type: string
        isActive:
          type: boolean
    Barber:
      type: object
      properties:
        id:
          type: string
          format: uuid
        publicName:
          type: string
        specialty:
          type: string
        photoUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
    BarberResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Barber"
        message:
          type: string
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        barberId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        startsAt:
          type: string
          example: "09:00"
        endsAt:
          type: string
          example: "09:30"
        status:
          $ref: "#/components/schemas/BookingStatus"
        priceSnapshot:
          type: number
          format: float
...
    MessageResponse:
      type: object
      properties:
        data:
          nullable: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string